backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==17.0.0
geopandas==1.0.1
shapely==2.0.4
folium==0.17.0
scikit-learn==1.5.1
python-dotenv==1.0.1
google-generativeai==0.7.2
beautifulsoup4==4.12.3
lxml==5.2.2